requirements_base.txt
Flask==3.1.2
Flask-Cors==6.0.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
gunicorn==22.0.0

SQLAlchemy==2.0.44
Flask-SQLAlchemy==3.1.1

psycopg2-binary==2.9.11

passlib[bcrypt]==1.7.4

python-dotenv==1.2.1

pytz==2025.2
tzdata==2025.2

httpx==0.28.1
sniffio==1.3.1
anyio==4.10.0

requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3

PyQRCode==1.2.1
pypng==0.20220715.0

cryptography==45.0.7
colorama==0.4.6
typing_extensions==4.14.1
click==8.2.1
MarkupSafe==3.0.2

openpyxl==3.1.5

# === NUEVAS DEPENDENCIAS ===
qrcode[pil]==7.4.2
reportlab==4.2.5
#pillow==11.0.0
#opencv-python-headless==4.10.0.84
#onnxruntime>=1.20.0
#insightface==0.7.3
#numpy==2.1.1